requirements.txt
python-dotenv
fastapi
fritzconnection
uvicorn